Address PR feedback
Add new tests to invokde executable dependency via a tool command.
