+++ test/Microsoft.DotNet.Tools.Tests.Utilities/Commands/DependencyToolInvokerCommand.cs
@@ -1,0 +1,30 @@
+// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using Microsoft.DotNet.Cli.Utils;
+using System.Runtime.InteropServices;
+using Microsoft.DotNet.ProjectModel;
+
+namespace Microsoft.DotNet.Tools.Test.Utilities
+{
+    public class DependencyToolInvokerCommand : TestCommand
+    {
+        public DependencyToolInvokerCommand()
+            : base("dotnet")
+        {
+        }
+
+        public CommandResult Execute(string commandName, string framework, string additionalArgs)
+        {
+            var args = $"dependency-tool-invoker {commandName} --framework {framework} {additionalArgs}";
+            return base.Execute(args);
+        }
+
+        public CommandResult ExecuteWithCapturedOutput(string commandName, string framework, string additionalArgs)
+        {
+            var args = $"dependency-tool-invoker {commandName} --framework {framework} {additionalArgs}";
+            return base.ExecuteWithCapturedOutput(args);
+        }
+    }
+}
